Title: 흥미로운 숫자, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 240802/흥미로운 숫자/interesting-numbers.java
@@ -2,8 +2,8 @@ import java.util.*;
 import java.io.*;
 
 public class Main {
-    static int Y;
-    static int X;
+    static long Y;
+    static long X;
     static int[] count = new int[10];
     static long answer = 0;
